GitDeveloper and GitPages to learn
Use GitHub’s suite of applications to learn collaborative coding.  Focus on Python and Data Engineering. Free tutorials and Exam questions to prepare for the Python institutes professional certification: The PCEP, The PCAP and PCPP 1. 
Provide your email and details to get access to a free subscription to an online learning platform to access free courses. Please specify which certificate you are interested in.